fix(abnormal): ensure float type for standard deviation calculation in PureAbnormalityDetector
- Updated the standard deviation score calculation in `PureAbnormalityDetector` to explicitly cast the result to float, enhancing type safety and preventing potential issues with type inference.
- This minor adjustment contributes to ongoing efforts to improve code reliability and maintainability within the abnormality detection module.

+++ src/brain_go_brrr/domain/abnormal/detector_pure.py
@@ -209,7 +209,7 @@ class PureAbnormalityDetector:
 
         # Confidence based on:
         # 1. Consistency (low std = high confidence)
-        std_score = 1.0 - min(np.std(scores_array) * 2, 1.0)
+        std_score = 1.0 - min(float(np.std(scores_array) * 2), 1.0)
 
         # 2. Extremity (far from 0.5 = high confidence)
         mean_score = np.mean(scores_array)
